refactor(mobile): give the album name and confirm dialogs one implementation
